+++ views/config/_menu-entry-fields.php
@@ -6,11 +6,11 @@
  * @author [Marc FARRE](https://marc.fun) for [CUZY.APP](https://www.cuzy.app)
  */
 
+use humhub\components\View;
 use humhub\modules\menuManager\models\Configuration;
 use humhub\modules\menuManager\models\MenuEntryConfig;
-use humhub\modules\ui\form\widgets\ActiveForm;
+use humhub\widgets\form\ActiveForm;
 use humhub\modules\ui\form\widgets\IconPicker;
-use humhub\modules\ui\view\components\View;
 
 
 /**
@@ -27,29 +27,29 @@ $MenuEntryConfigAttributeHints = (new MenuEntryConfig())->attributeHints();
 
 <?php if ($configurationAttributeLabels[$attribute] ?? null) : ?>
 
-    <div class="alert alert-default">
+    <div class="alert alert-light" role="alert">
         <h4 style="margin-bottom: 15px;"><strong><?= $configurationAttributeLabels[$attribute] ?></strong></h4>
 
         <div class="row">
-            <div class="col-md-3">
+            <div class="col-lg-3">
                 <?= $form->field($model, $attribute . '[displayState]')
                     ->dropDownList(MenuEntryConfig::getDisplayStateLabels())
                     ->label($MenuEntryConfigAttributeLabels['displayState'] ?? '')
                     ->hint($MenuEntryConfigAttributeHints['displayState'] ?? '') ?>
             </div>
-            <div class="col-md-3">
+            <div class="col-lg-3">
                 <?= $form->field($model, $attribute . '[icon]')
                     ->widget(IconPicker::class)
                     ->label($MenuEntryConfigAttributeLabels['icon'] ?? '')
                     ->hint($MenuEntryConfigAttributeHints['icon'] ?? '') ?>
             </div>
-            <div class="col-md-3">
+            <div class="col-lg-3">
                 <?= $form->field($model, $attribute . '[label]')
                     ->textInput()
                     ->label($MenuEntryConfigAttributeLabels['label'] ?? '')
                     ->hint($MenuEntryConfigAttributeHints['label'] ?? '') ?>
             </div>
-            <div class="col-md-3">
+            <div class="col-lg-3">
                 <?= $form->field($model, $attribute . '[sortOrder]')
                     ->textInput(['type' => 'number', 'step' => 1, 'min' => 1, 'max' => 10000])
                     ->label($MenuEntryConfigAttributeLabels['sortOrder'] ?? '')