+++ views/config/index.php
@@ -6,11 +6,11 @@
  * @author [Marc FARRE](https://marc.fun) for [CUZY.APP](https://www.cuzy.app)
  */
 
+use humhub\components\View;
 use humhub\modules\menuManager\models\Configuration;
 use humhub\modules\menuManager\Module;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\modules\ui\view\components\View;
-use humhub\widgets\Button;
+use humhub\widgets\bootstrap\Button;
+use humhub\widgets\form\ActiveForm;
 
 
 /**
@@ -27,13 +27,13 @@ $module = Yii::$app->getModule('menu-manager');
     <div class="panel-heading">
         <strong><?= $module->getName() ?></strong>
 
-        <div class="help-block">
+        <div class="text-body-secondary">
             <?= $module->getDescription() ?>
         </div>
     </div>
 
     <div class="panel-body">
-        <div class="alert alert-info cuzy-free-module-info">
+        <div class="alert alert-info cuzy-free-module-info" role="alert">
             This module was created and is maintained by
             <a href="https://www.cuzy.app/"
                target="_blank">CUZY.APP (view other modules)</a>.